requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
feedparser>=6.0,<7.0
beautifulsoup4>=4.12,<5.0
tweepy>=4.14,<5.0
celery>=5.3,<6.0
redis>=4.6,<5.0
schedule>=1.2,<2.0
textblob>=0.17,<1.0
pandas>=2.1,<3.0
numpy>=1.25,<2.0
matplotlib>=3.7,<4.0
seaborn>=0.12,<1.0
plotly>=5.15,<6.0
django-celery-beat>=2.5,<3.0
gunicorn>=21.2,<22.0
whitenoise>=6.5,<7.0
django-filter>=25.1,<26.02
weasyprint>=61.0,<62.0
reportlab>=4.0,<5.0